# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KnowledgeMind RAG API",
    "description": "企业级智能知识库 RAG 系统 - DeepSeek API 驱动",
    "version": "1.1.0"
  },
  "paths": {
    "/api/v1/chat": {
      "post": {
        "tags": [
          "Chat"
        ],
        "summary": "Chat",
        "operationId": "chat_api_v1_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chat/stream": {
      "post": {
        "tags": [
          "Chat"
        ],
        "summary": "Chat Stream",
        "operationId": "chat_stream_api_v1_chat_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/search": {
      "post": {
        "tags": [
          "Chat"
        ],
        "summary": "Search",
        "operationId": "search_api_v1_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/classify": {
      "post": {
        "tags": [
          "Chat"
        ],
        "summary": "Classify Query",
        "operationId": "classify_query_api_v1_classify_post",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Query"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/conversations/{conversation_id}": {
      "get": {
        "tags": [
          "Chat"
        ],
        "summary": "Get Conversation",
        "operationId": "get_conversation_api_v1_conversations__conversation_id__get",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conversation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Chat"
        ],
        "summary": "Delete Conversation",
        "operationId": "delete_conversation_api_v1_conversations__conversation_id__delete",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conversation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationDeleteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatRequest": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 10000,
            "minLength": 1,
            "title": "Query",
            "description": "用户问题"
          },
          "conversation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          },
          "reasoning_mode": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReasoningMode"
              },
              {
                "type": "null"
              }
            ]
          },
          "top_k": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 20.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Top K",
            "default": 5
          },
          "temperature": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 2.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature",
            "default": 0.1
          },
          "use_fallback": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Use Fallback"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "ChatRequest"
      },
      "ChatResponse": {
        "properties": {
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "segments": {
            "items": {
              "$ref": "#/components/schemas/ChatSegment"
            },
            "type": "array",
            "title": "Segments",
            "default": []
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "type": "array",
            "title": "Citations",
            "default": []
          },
          "reasoning_content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasoning Content"
          },
          "confidence_score": {
            "type": "number",
            "title": "Confidence Score",
            "default": 0.0
          },
          "conversation_id": {
            "type": "string",
            "title": "Conversation Id"
          },
          "token_usage": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token Usage"
          },
          "model_used": {
            "type": "string",
            "title": "Model Used",
            "default": ""
          },
          "reasoning_mode_used": {
            "type": "string",
            "title": "Reasoning Mode Used",
            "default": ""
          },
          "is_fallback": {
            "type": "boolean",
            "title": "Is Fallback",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "answer",
          "conversation_id"
        ],
        "title": "ChatResponse"
      },
      "ChatSegment": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "type": "array",
            "title": "Citations",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "ChatSegment"
      },
      "Citation": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "document_title": {
            "type": "string",
            "title": "Document Title"
          },
          "page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page"
          },
          "paragraph": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paragraph"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "score": {
            "type": "number",
            "title": "Score"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "document_title",
          "content",
          "score"
        ],
        "title": "Citation"
      },
      "ConversationDeleteResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "conversation_id": {
            "type": "string",
            "title": "Conversation Id"
          }
        },
        "type": "object",
        "required": [
          "success",
          "conversation_id"
        ],
        "title": "ConversationDeleteResponse"
      },
      "ConversationDetail": {
        "properties": {
          "conversation_id": {
            "type": "string",
            "title": "Conversation Id"
          },
          "created_at": {
            "type": "number",
            "title": "Created At"
          },
          "turns": {
            "items": {
              "$ref": "#/components/schemas/ConversationTurn"
            },
            "type": "array",
            "title": "Turns"
          }
        },
        "type": "object",
        "required": [
          "conversation_id",
          "created_at",
          "turns"
        ],
        "title": "ConversationDetail"
      },
...
      "ConversationTurn": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "timestamp": {
            "type": "number",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "query",
          "answer",
          "timestamp"
        ],
        "title": "ConversationTurn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ReasoningMode": {
        "type": "string",
        "enum": [
          "non_think",
          "think_high",
          "think_max"
        ],
        "title": "ReasoningMode"
      },
      "SearchRequest": {
        "properties": {
          "query": {
            "type": "string",
            "minLength": 1,
            "title": "Query"
          },
          "top_k": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Top K",
            "default": 10
          },
          "filters": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filters"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "SearchRequest"
      },
      "SearchResponse": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/SearchResult"
            },
            "type": "array",
            "title": "Results"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "query": {
            "type": "string",
            "title": "Query"
          }
        },
        "type": "object",
        "required": [
          "results",
          "total",
          "query"
        ],
        "title": "SearchResponse"
      },
      "SearchResult": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "content",
          "score",
          "metadata"
        ],
        "title": "SearchResult"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}